Decouple cstg
