+++ karma.conf.maker.js
@@ -111,6 +111,8 @@ module.exports = function(codeCoverage, browserstack, watchMode, file, disableFe
   var plugins = newPluginsArray(browserstack);
 
   var files = file ? ['test/test_deps.js', file, 'test/helpers/hookSetup.js'].flatMap(f => f) : ['test/test_index.js'];
+  // var fileArray = Array.isArray(file) ? file : [file];
+  // var files = file ? ['test/test_deps.js', ...fileArray, 'test/helpers/hookSetup.js'].flatMap(f => f) : ['test/test_index.js'];
   // This file opens the /debug.html tab automatically.
   // It has no real value unless you're running --watch, and intend to do some debugging in the browser.
   if (watchMode) {
